Reworked Setlist schema
I needed a way to properly link the song data to the Songs collection, which hadn't come up as a problem until working on how to edit the Setlist request and wanting to change the position. Changed the add view for Setlist to include a hidden input containing the song ID. Also changed the song title and artist fields to readonly so the user can't muck with them.